productor/scripts/requirements.txt
ultralytics
opencv-python  # ← Corrige esto (no "nopencv-python")
kafka-python
torch